requirements.txt
numpy==1.15.4
torchvision==0.2.1
requests==2.20.1
opencv_contrib_python==3.4.4.19
torch==0.4.1
Pillow==5.3.0
scikit_learn==0.20.2
